+++ uspto/peds/tasks.py
@@ -2,7 +2,7 @@
 # (c) 2017 Andreas Motl <andreas@ip-tools.org>
 import celery
 from uspto.peds.client import UsptoPatentExaminationDataSystemClient
-from uspto.util.tasks import GenericDownloadTask
+from uspto.util.tasks import GenericDownloadTask, AsynchronousDownloader
 
 class UsptoPatentExaminationDataSystemDownloadTask(GenericDownloadTask):
     name = 'uspto.peds.tasks.UsptoPatentExaminationDataSystemDownloadTask'
@@ -15,3 +15,6 @@ def download_task(self, query, options=None):
     http://docs.celeryproject.org/en/latest/whatsnew-4.0.html#the-task-base-class-no-longer-automatically-register-tasks
     """
     return self.process(query, options)
+
+class UsptoPatentExaminationDataSystemDownloader(AsynchronousDownloader):
+    task_function = download_task